+++ src/GameServer/KcpSharp/KcpSocketTransportForRawChannel.cs
@@ -6,14 +6,14 @@ namespace KcpSharp
 {
     internal sealed class KcpSocketTransportForRawChannel : KcpSocketTransport<KcpRawChannel>, IKcpTransport<KcpRawChannel>
     {
-        private readonly int? _conversationId;
+        private readonly long? _conversationId;
         private readonly KcpRawChannelOptions? _options;
 
         private Func<Exception, IKcpTransport<KcpRawChannel>, object?, bool>? _exceptionHandler;
         private object? _exceptionHandlerState;
 
 
-        internal KcpSocketTransportForRawChannel(Socket socket, EndPoint endPoint, int? conversationId, KcpRawChannelOptions? options)
+        internal KcpSocketTransportForRawChannel(Socket socket, EndPoint endPoint, long? conversationId, KcpRawChannelOptions? options)
             : base(socket, endPoint, options?.Mtu ?? KcpConversationOptions.MtuDefaultValue)
         {
             _conversationId = conversationId;
